Convert firepower modifiers to integer percentages.

+++ OpenRA.Game/Traits/TraitsInterfaces.cs
@@ -157,7 +157,7 @@ namespace OpenRA.Traits
 	public interface IRenderModifier { IEnumerable<IRenderable> ModifyRender(Actor self, WorldRenderer wr, IEnumerable<IRenderable> r); }
 	public interface IDamageModifier { int GetDamageModifier(Actor attacker, DamageWarhead warhead); }
 	public interface ISpeedModifier { int GetSpeedModifier(); }
-	public interface IFirepowerModifier { float GetFirepowerModifier(); }
+	public interface IFirepowerModifier { int GetFirepowerModifier(); }
 	public interface ILoadsPalettes { void LoadPalettes(WorldRenderer wr); }
 	public interface IPaletteModifier { void AdjustPalette(IReadOnlyDictionary<string, MutablePalette> b); }
 	public interface IPips { IEnumerable<PipType> GetPips(Actor self); }
